Add sub launch pad and modify the rest accordingly
We've added a stub for a polkadot-wiki style launch pad. Accordingly,
we've moved Getting Started into the Documentation section. The links to
the documentation have been removed from the footer. I think they are
superfluous.

As a result, the footer is a bit empty now. Maybe something can be
filled in by the community.